Using a GA to find a solution to the traveling salesman problem (TSP).
Problem Statement:

“Given a list of cities and the distances between each pair of cities, what is the shortest possible route that visits each city and returns to the origin city”

Genetic Algorithm:

Genetic algorithm is started with a set of solutions (represented by chromosomes) called population. Solutions from one population are taken and used to form a new population.The new population will be better than the old one. Solutions which are selected to form new solutions (offspring) are selected according to their fitness, the more suitable they are the more chances they have to reproduce. This is repeated until some condition for example number of populations or improvement of the best solution is satisfied. Genetic Algorithm is a paradigm that has proved to be a unique approach for solving various mathematical problems which other gradient type of mathematical optimizers have failed to reach,Ant colony optimization has been applied successfully to a large number of difficult combinatorial optimization problems.